+++ health_check/backends.py
@@ -5,38 +5,13 @@ import logging
 
 from django.utils.six import text_type
 from django.utils.translation import ugettext_lazy as _
-from django.utils.encoding import python_2_unicode_compatible
+
+from health_check.exceptions import HealthCheckException
 
 logger = logging.getLogger('health-check')
 
 
-class HealthCheckStatusType(object):
-    unavailable = 0
-    working = 1
-    unexpected_result = 2
-
-
-@python_2_unicode_compatible
-class HealthCheckException(Exception):
-    type = _("unknown error")
-
-    def __init__(self, message):
-        self.message = message
-
-    def __str__(self):
-        return "%s: %s" % (self.type, self.message)
-
-
-class ServiceUnavailable(HealthCheckException):
-    type = _("unavailable")
-
-
-class ServiceReturnedUnexpectedResult(HealthCheckException):
-    type = _("unexpected result")
-
-
 class BaseHealthCheckBackend(object):
-
     def __init__(self):
         self.errors = []
