+++ health_check/plugins.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 # This is heavily inspired by the django admin sites.py
 
-from health_check.backends.base import BaseHealthCheckBackend
+from health_check.backends import BaseHealthCheckBackend
 
 
 class AlreadyRegistered(Exception):
